pyproject.toml
[project]
name = "postgres-supabase-iac"
version = "0.1.0"
description = ""
authors = [
    {name = "Giorgio Segalla",email = "giorgiosgl1997@gmail.com"}
]

[tool.poetry.scripts]
postgres-supabase = "postgres_supabase_iac.__main__:app"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "^3.11"
sqlmodel = ">=0.0.24 <0.0.25"
typer = ">=0.15.4 <0.16.0"
python-decouple = "^3.8"
psycopg2-binary = "^2.9.9"

[tool.poetry.group.dev.dependencies]
ipython = "^9.2.0"
pytest = "^8.3.5"